requirements.txt
ijson
pandas
numpy
matplotlib
tqdm
openai
zhipuai
pillow